refactor: move and rename AboutUsSummaryCard into GenericPagesComponents
