+++ src/pages/about-us/leadership.tsx
@@ -2,7 +2,7 @@ import { NextPage, GetStaticProps, GetStaticPropsResult } from "next";
 
 import CMSClient from "@/node-lib/cms";
 import ContactCard from "@/components/GenericPagesComponents/ContactCard";
-import AboutUsSummaryCard from "@/components/pages/AboutUs/AboutUsSummaryCard";
+import GenericSummaryCard from "@/components/GenericPagesComponents/GenericSummaryCard";
 import BioCardList from "@/components/GenericPagesComponents/BioCardList";
 import AboutIntroCard from "@/components/ArchivedComponents/AboutIntoCard/AboutIntroCard";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
@@ -22,7 +22,7 @@ const AboutUsLeadership: NextPage<AboutPageProps> = ({ pageData }) => {
   return (
     <Layout seoProps={getSeoProps(seo)} $background={"white"}>
       <MaxWidth $mb={[56, 80]} $pt={[64, 80]}>
-        <AboutUsSummaryCard {...pageData} />
+        <GenericSummaryCard {...pageData} />
         <AboutIntroCard
           image={{
             illustration: "supporting",
